Global functions registerd for backend are moved to 'main.ts'.
