+++ web/src/views/HomeView.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { onMounted, useTemplateRef } from 'vue';
+import { useTemplateRef } from 'vue';
 
 import FormArea from '@/components/FormArea.vue';
 import OperationArea from '@/components/OperationArea.vue';
@@ -8,7 +8,6 @@ import { useLogStore } from '@/store/log';
 import { useMediaDataStore } from '@/store/media-data';
 
 import { formItemInfo } from '@/utils/form-item-info';
-import { bytesToSize } from '@/utils/show';
 
 const form = useTemplateRef('form');
 
@@ -29,7 +28,7 @@ function handleFormSubmit(action: string) {
         ...form.value.data,
     };
 
-    logMessage(`Request: ${JSON.stringify(data, null, 2)}`);
+    log.log(`Request: ${JSON.stringify(data, null, 2)}`);
 
     // Backend call
     webui.submitUrl(JSON.stringify(data)).then((response) => {
@@ -38,43 +37,6 @@ function handleFormSubmit(action: string) {
         }
     });
 }
-
-function logMessage(message: string) {
-    log.log(message);
-}
-
-function showDownloadProgress(rawData: Uint8Array) {
-    const data = new TextDecoder().decode(rawData);
-    const json = JSON.parse(data);
-
-    const filename = json.filename;
-    const downloaded_bytes = bytesToSize(json.downloaded_bytes);
-    const total_bytes = bytesToSize(json.total_bytes);
-    const speed = bytesToSize(json.speed);
-    const progress = (json.downloaded_bytes / json.total_bytes) * 100;
-    logMessage(
-        `Downloading ${filename}: ${downloaded_bytes} / ${total_bytes} (${progress.toFixed(2)}%) Speed: ${speed}/s`,
-    );
-}
-
-function showDownloadInfo(rawData: Uint8Array) {
-    const data = new TextDecoder().decode(rawData);
-    logMessage(data);
-}
-
-declare global {
-    interface Window {
-        logMessage: (message: string) => void;
-        showDownloadProgress: (rawData: Uint8Array) => void;
-        showDownloadInfo: (rawData: Uint8Array) => void;
-    }
-}
-
-onMounted(() => {
-    window.logMessage = logMessage;
-    window.showDownloadProgress = showDownloadProgress;
-    window.showDownloadInfo = showDownloadInfo;
-});
 </script>
 
 <template>
